+++ src/Cooked.hs
@@ -10,5 +10,6 @@ import Cooked.MockChain as X
 import Cooked.Pretty as X
 import Cooked.ShowBS as X
 import Cooked.Skeleton as X
+import Cooked.Staged as X
 import Cooked.Tweak as X
 import Cooked.Wallet as X
